MOD modal de cerrar sesion
