+++ resources/js/views/Home.vue
@@ -37,41 +37,10 @@
     <a class="scroll-to-top rounded" href="#page-top">
         <i class="fas fa-angle-up"></i>
     </a>
-    <!-- Logout Modal-->
-    <div class="modal fade" id="logoutModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
-        aria-hidden="true">
-        <div class="modal-dialog" role="document">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="exampleModalLabel">Ready to Leave?</h5>
-                    <button class="close" type="button" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">×</span>
-                    </button>
-                </div>
-                <div class="modal-body">Select "Logout" below if you are ready to end your current session.</div>
-                <div class="modal-footer">
-                    <button class="btn btn-secondary" type="button" data-dismiss="modal">Cancel</button>
-                    <button class="btn btn-primary" @click="logout">Logout</button>
-                </div>
-            </div>
-        </div>
-    </div>
 </template>
 
 <script setup>
-    import { ref } from 'vue'
-    import { useRouter } from 'vue-router'
     import Sidebar from '../components/Sidebar.vue'
     import Navbar from '../components/Navbar.vue'
     import BreadCrumb from '../components/Breadcrumb.vue'
-
-    // Router
-    const router = useRouter()
-
-    // Métodos
-    function logout() {
-         $('#logoutModal').modal('hide')
-        localStorage.removeItem("auth_token")
-        router.push('/login')
-    }
 </script>
